requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
Flask==2.2.5
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
python-dotenv==0.21.1
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.3
Werkzeug==2.2.3
yarl==1.9.2
zope.interface==6.0
gunicorn==20.1.0
flask-cors==3.0.10

